Fix double

+++ cpp/oneapi/dal/algo/decision_forest/backend/gpu/train_splitter_impl_dpc.cpp
@@ -484,7 +484,7 @@ sycl::event train_splitter_impl<Float, Bin, Index, Task>::best_split(
                     // Regression case
                     const Index work_size = local_size / act_bin_block;
                     Index count = 0;
-                    double sum = 0;
+                    Float sum = 0;
                     Float weight = 0;
                     const Index bin_id = local_id % act_bin_block;
                     const Index loc_bin_pos = bin_id * hist_prop_count;
@@ -525,7 +525,7 @@ sycl::event train_splitter_impl<Float, Bin, Index, Task>::best_split(
                     }
                     // Finalize regression case by calculating MSE
                     item.barrier(sycl::access::fence_space::local_space);
-                    double mse = 0;
+                    Float mse = 0;
                     const Float mean = local_hist[loc_bin_pos + 1] / local_hist[loc_bin_pos + 0];
                     for (Index row_idx = local_id / act_bin_block; row_idx < row_count;
                          row_idx += work_size) {
